Split _compute/runtime_test.py into four co-located test modules
runtime_test.py named a module that no longer exists: runtime.py was
deleted when its contents were extracted into layouts.py, ions.py,
bindings.py and state.py. Split its 3,200 lines and 111 tests to match,
so each foo.py again has a sibling foo_test.py.

All 111 test bodies move byte-identically; only the enclosing class and
file change. CellRuntimeStateTest's 66 methods split four ways by
subject -- 12 layout, 28 ion, 17 binding, 9 residual state -- and the
other 12 classes move whole. Final counts are 32 / 28 / 24 / 27.

The three module-level fixtures shared across the split go into a new
private _testing.py. This is required for correctness, not tidiness:
_RuntimeTestTwoOwnerChannel is @register_channel-decorated, and
mech/_registry.py raises ValueError on duplicate registration, so
copying it into each file that needs it would kill the suite at
collection time. It is now defined exactly once, and its __module__ is
repointed at its new home. The clamp stubs have a single consumer, so
they are hoisted to the top of layouts_test.py instead.

Each new file imports only what its own tests use rather than the union
of the old import block; the three mid-file module-level import sites
are hoisted into whichever file they landed in. Imports that sit inside
a test method body are left where they are, since moving them would
edit a body.

Verified by test-name fingerprint: the sorted collection of test names
under _compute/ is unchanged, which a passing count alone cannot show.
141 passed in _compute/, 2240 passed and 19 skipped repo-wide, both
matching the pre-split baseline.
